monitor: use signed value for factor and offset
The code was using size_t for factor and offset. For some fans the
offset could be negative, so it needs to use signed types.

This commit uses int64_t for both factor and offset to make it
consistent, although factor should really be an unsigned value.

+++ monitor/tach_sensor.hpp
@@ -88,8 +88,8 @@ class TachSensor
                    bool hasTarget,
                    size_t funcDelay,
                    const std::string& interface,
-                   size_t factor,
-                   size_t offset,
+                   int64_t factor,
+                   int64_t offset,
                    size_t timeout,
                    const sdeventplus::Event& event);
 
@@ -125,7 +125,7 @@ class TachSensor
         /**
          * @brief Returns the factor of the sensor target
          */
-        inline size_t getFactor() const
+        inline int64_t getFactor() const
         {
             return _factor;
         }
@@ -133,7 +133,7 @@ class TachSensor
         /**
          * @brief Returns the offset of the sensor target
          */
-        inline size_t getOffset() const
+        inline int64_t getOffset() const
         {
             return _offset;
         }
@@ -272,12 +272,12 @@ class TachSensor
         /**
          * @brief The factor of target to get fan rpm
          */
-        const size_t _factor;
+        const int64_t _factor;
 
         /**
          * @brief The offset of target to get fan rpm
          */
-        const size_t _offset;
+        const int64_t _offset;
 
         /**
          * @brief The input speed, from the Value dbus property
